+++ policyengine_us/variables/household/healthcare_benefit_value.py
@@ -23,7 +23,11 @@ class healthcare_benefit_value(Variable):
         "co_omnisalud",
         "assigned_co_premium_assistance",
         "ct_covered_connecticut",
+        "ma_connector_care",
         "md_premium_assistance",
+        "nj_njhps",
         "assigned_nm_premium_assistance",
         "or_healthier_oregon_cost",
+        "vt_premium_assistance",
+        "wa_cascade_care_savings",
     ]